refactors code, fixes some styling errors
