+++ src/navigation/RootNavigator.tsx
@@ -1,7 +1,6 @@
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
-import NavigationBar from '../components/NavigationBar';
-import HomeScreen from '../screens/HomeScreen/Home';
+import NavigationBar from './NavigationBar';
 import PlayScreen from '../screens/PlayScreen/Play';
 
 const Stack = createStackNavigator();
@@ -9,9 +8,8 @@ const Stack = createStackNavigator();
 export default function RootNavigation() {
   return (
     <NavigationContainer>
-      <Stack.Navigator>
+      <Stack.Navigator screenOptions={{headerShown: false}}>
         <Stack.Screen name="Tabs" component={NavigationBar} />
-        <Stack.Screen name="Audio" component={HomeScreen} />
         <Stack.Screen name="Play" component={PlayScreen} />
       </Stack.Navigator>
     </NavigationContainer>
